TaskList: Modify to use Streams in implementation
As per the requirment in streams, the TaskList class has been modified
to use streams in its implementation. Using streams also gives shorter
code and better readability.

Most of the for loops were replaced with IntStream in the TaskList class.
It is probably the easiest way to convert a for loop to a stream.

No real impact on performance was observed. The chatbot should still be
able function as per normal.