Add assertions into key parts of the code
The code does not have any assertions statements within at this moment.

Assertions in Java are useful in order to determine whether the code
is working as intended or not.

Let's add in assertions into the existing code in order to ensure that
the code is working as intended at different points within the code.

The assertion statements are being added at key points to ensure that
the application is working as intended and can carry out what the task
it needs to do.