Add assert statements to document assumptions
The code makes some assumptions at various places.

This might make regressions and bugs harder to spot as we continue
to develop the code.

Let's include assertion checks at these places where assumptions are
made.

This allows us to check assumptions made in the code and ensure that
such assumptions are still being upheld, even as we change the code.