Improve Code Quality
Currently, the naming of methods in the Ui class are misleading and the conditional statements in the Parser class are hard to read.

This can cause others to have a hard time reading the code.

Let's improve the naming conventions and use switch case statements instead!

This makes it easier for others to read the code.