Add assertions and reorganise exception handling
Exception handling for some methods accessing the chatbot TaskList
are disorganised and spread across various classes (eg. Chatbot, TaskList).

The throwing of namely the InvalidTaskNumberException in different methods is
messy and reduces code readability.

Organising that exception handling will adhere more to OOP and will be less messy.

Let's move the checking of the task number input to the TaskList class and have the methods
there throw the InvalidTaskNumberException. Let's add assertions in the classes to assert that
subsequent calls to methods in the program flow will have valid inputs after handling this
exception in the TaskList methods.