Add error handling
Assertions are used to define assumptions about the program state so that the runtime can verify them. Exceptions are used to check if a user's input fall within a set of expected inputs and allows the program to react accordingly if it does not.

The Storage class assumes that all the Tasks stored in the data file has the indentifier "T", "D" or "E". Adding an assertion to check if this is true ensures that the program runs as desired.

The Parser class parse given inputs and call the corresponding methods to execute the specified commands. Adding an exception to check if commands that act on integers are indeed followed by digits.