Extract createTask method from Storage.load
Storage::load has a long and complicated method body

This affects the readability of the code.

Let's extract out the portion of code relating to creating the task into
a new function, createTask.

Let's further extract the portion of the code that deals with
determining if the task has been done into a new function, getTaskStatus.

This reduces the size of each method, which makes the code easier to
read.