Change find date implementation
There are multiple hashtables keeping track of different types of tasks with different sets of dates.

This is unnecessariliy complicated, as trying to find different types of tasks on different dates requires multiple commands.

Let's combine all of the tables into one single hashtable to keep track of the dates.

This is so that a date can query for all different types of tasks using just one command, which will make the chatbot easier to use.