Add find free dates functionality
There is no way for the user to find what are the next available dates in his/her schedule.

If the user already has an event on a certain date, he/she might not want to add another event on the same date.

Let's add a functionality for the user to find what are the next available dates with the command "freedates", which will return the 5 closest date to the current date that is free.