Add various assert statements
Ensured the parameters passed to initialising the various tasks are not
empty. Ensured the date format is correct for Deadline and Event.
Ensured dialogContainer only has one child when initialised.

Removed a bug that causes errors to not show up immediately in the Duke
chatbot and instead wait for valid input from user before showing all
errors simultaneously.