Use java assertions to detect potential bugs
The code does not have any assertions to validate some assumptions
made in the class TaskList and Storage.

This makes the code vulnerable to potential bugs and runtime
errors if the assumptions do not hold after some code change.

Let's validate the instantiation of the Storage class indeed
creates a valid file for data storage and the load method
can load the encoded tasks correctly.

Let's also make sure the description passed to the add
method of TaskList is indeed non-empty.

Doing so allows the two class to be less error-proned,
and we can be alerted to any errors if the assertions
fail.