Removal of dead code and static keywords for Jarvis, Parser and Ui classes.
Dead code and unused variables are redundant and confuses other developers.
Wrong use of static keyword confuses others developers as class level methods and attributes are not actually meant to be as such.

Delete dead code and remove unnecessary static keywords.

So as to declutter code and fix wrong class level methods.