Add a tag to the list command
Currently, users only have a list commmand to list all the tasks.

To create reminders for upcoming tasks, users need to be able to view
the tasks that are due before/within the time period.

A flag can be used at the end of the list command of the format:
'list flag'. The time period for both deadline and event starts at the
start of the day at 00:00. The end is determined by the flag indicated
by the user.

Let's add the following flags:
* 'today' : lists all tasks that in the period today 00:00 to tmr 00:00
* 'week' : lists all tasks that are in the period today 00:00 to one
week later 00:00
* 'fort': lists all tasks that are in the period today 00:00 to 2 weeks
later 00:00