Fix mass operations feature
Due to errors in the code which went undiscovered in past commits, the feature to mass mark, unmark and delete raised runtime exceptions everytime it was called.

The issues lie in the parsing of the command input and not throwing an exception for unexpected inputs, which are fixed in this commit.