Add asserts to Commands and TaskList
Commands and Tasklist have made some assumptions in order for the method to function properly
(e.g. TaskList.add(index i) requires i to be more than 0 and less than the list.size()).

Adding assertions can document important assumptions that should hold for this methods to work as intended.

Let's add assertions to various points in the code to document valid assumptions that are implicit and make it as clear as possible for other developers.