Add assertions to code
Add assertions to code

This project contains many classes that work together.

These classes are prone to errors.

Let's add assertions to some parts of the code to:
- indicate possible bugs
- ensure correctness of code

If the runtime detects an assertion failure, it typically takes some
drastic action such as terminating the execution with an error message.