Implement sort feature
Sort feature is implemented for C-Sort extension.

New commands are added into chatbot to sort deadlines and events using
the "sort deadlines" and "sort events" commands respectively.

Instead of printing the message string in the TaskList methods, it is now printed
from the methods in the Ui class.

It makes more sense for the Ui to return message strings instead of the
TaskList class since the TaskList class handles the operations on the ArrayLists
of Tasks.

The error handling for sort commands are not yet implemented.