Add assertions to some classes
In classes Todos, Deadlines and Events, assertions are added to document assumptions like descriptions, date or time should not be nul.

In other classes, assertions are added to document some important assumptions that should hold at various points in the code. An example is in Duke.class, user input now cannot be null when clicking the send button.

Using assert statements wisely can lead to more maintainable and robust code. However, as with all tools, it's important to use them judiciously and in the appropriate context.