Add Assertions for Code Reliabilty
In this commit, I have added a series of assert statements to increase
the reliability and robustness of the project.

Assertions help the programmers to catch bugs in early stages making it
easier for the programmers to fix the issue.

Key Changes:
-> Added assertions to TaskList and parser.
-> Checked for edge cases to ensure proper correctness of the logic.
-> Improved error handling by using assertion to prevent haywire
behaviour of the program.
More documentation on assertions please refer -> https://www.javatpoint.com/assertion-in-java