Added Assertion Statements
There exist certain assumptions made about the arguments being
passed into some methods.

It is important that we add assertion to check for possible bugs and
terminate the execution so that it is more safe.

Assertion statements have been used within the method to check for certain
characteristics of the arguments (eg. String input is not null)