Implement A-Assertions
There is no usage of assertions in all classes.

The program is unable to check for different assumptions that can cause
runtime error.

Using assertions in different part of the programs allows us to test
for bug when dealing with different inputs that users can potentially
enter.

Java assertions is preferable because it can flag out assertions that
fails but will not terminate the program.