Improve Code Quality
Improve Code Quality

The code is messy and complicated.

Making the code more readable is crucial for other developers to
understand the code, and ensure that we will have greater clarity with
the code in the future.

Let's improve code quality by avoiding long methods, complicated
expressions and practice KISSing.

Doing so makes the code easier on the eyes to see and understand.