Fix bugs related to GUI commit
There were several error messages that were not being displayed on the
GUI by Jeeves.

Not having those messages be displayed to the user greatly affects user
experience as they would not have any clue what or if anything was wrong
at all.

Let's update the code such that all error messages are properly shown
by Jeeves and no part of the code attempts to print to a command line
output.