Add ability to archive task list
Users are able to store the task list on their computers.

However, they are not able to start the application on a clean slate
with an empty task list unless they delete everything in the task list.

Let's add an archive file so that users can choose to store their
task lists temporarily and use the application without past tasks.

Improvements to this feature can be made in the future.