Implement A-CodeQuality
Codes in some classes have borderline bad code quality.

Readability is not the priority at that point in time. As more features
are added, readability of the code will be beneficial to not only the
developers but to others reading it.

Refactor to improve code quality such as adhering to SLAP, shortening
long methods and structuring the code logically.

To achieve readability and understandability, these improvement of code
quality aspects will allow others to easily understand the codes.