Add help functionality to the application
When a user opens the application, it does not provide any prompts or
guide on how the application can be used. As such, the application
may seem intimidating and unfriendly to new users, or users who
have forgotten the commands.

Let's add,
* add a welcome message to make the user feel more welcome, as
well as introduce the help function to the user
* add a help function to the application to allow the user
to see all commands that the application has to offer

Doing this is important in order to allow new users to learn how to
use the application, while providing existing users a refresher
in case they have forgotten the commands needed to use the application.