Add functionality to detect duplicates
Chatbot did not recognise and handle duplicate entries of the same task. This
led to clunky UX as users added multiple entries of the same task.

Added functionality to inform users of duplicate entries when adding new tasks, to list
duplicates, and to delete all duplicates if any.

Fixed a bug regarding delete where the wrong number of remaining tasks left is shown.