Clean code quality using checkstyle as a guide
Code quality is currently not clean. checkstyle shows that there
are errors in the styling of my code.

Additional space are being removed, as well as unnecessary code.
Checkstyle guide is used as reference.

These are done as then there would be a check list of styling
that i can check against for any errors.