Generate a new JAR file
The current JAR file is based on the old version of the code, without
reflecting changes made to the code that include incorporating
JavaFX GUI.

The current JAR file is thus no longer suitable as the latest release.

Let's generate a new JAR file that reflects the latest changes,
which include incorporating JavaFX GUI.

This new JAR file will henceforth be used as the latest release.