Add Code Quality Improvements.
In Parser.java, the happy path is nested, we use guard clauses to make
it clearer and more prominent.
There are also multiple levels of abstraction within the switch
statements, which make the method parseCommand excessively long.
As such, we abstract out string processing statements
into smaller and shorter methods.

In Ui.java,
* We remove dead code that print to the terminal since the text UI is
no longer used.
* We avoid magic strings for the farewell text as it is used to check
for an exit command in MainWindow.java

We also make other improvements and minor adjustments to tests and
commands, and delete dead code in this refactoring.