Include Java assertions in program code
Currently, there are no Java assertions in the program code.

It is hard to keep track of important assumptions in various parts of
the program as a result, so future regressions may be more easily
resolved.

Let's add Java assertions to the program code.

This will make it easier to track important assumptions in various
parts of the program and thus resolve future regressions.