Add assertions to Ducky and TaskList
There is a possibility that the file path given to Ducky or a task added
to the TaskList is empty or null respectively.

Making sure invariants for both cases are held improves safety especially
if future commits allow file path or format to be changed.

Let's add assertions for both cases.