Abstract the conversion of a task to its data string to be stored in the data text file.
Storage saveCurrentTask() method gets the task description and boolean about whether it is done and converts them into data string to be stored in data text file.

There are multiple levels of abstraction in the saveCurrentTask() method  making it difficult to read. There is also a lot of code duplication.

Creating a new abstract method toDataString() into the Task class ensures that all instances of classes that inherit from Task class will have its own adoption of a toDataString() method.

Let's create an abstract method in Task class called toDataString() and add the toDataString() method adaptation into all classes that inherit from Task, such as Deadline, Todo and Event.

Let's reduce the code duplication in the saveCurrentTask() method in the Storage class and remove the if else statements that differentiate the different types of tasks using instanceof operator.