Improve readability of all methods in the Parser class.
1. convertInputDateAndTimeIntoLocalDateTime()

Year is represented by Integer.parseInt(dayMonthYearArray[2]) without any indication that it is equal to year. Same with month, day, hours and minutes.

convertInputDataAndTimeIntoLocalDateTime() method is difficult to read because of Integer.parseInt(dayMonthYearArray[i]). It is not clear what each element in dayMonthYearArray refers to immediately. Same with month, day, hours and minutes.

Make the code look simpler by assigning a variable year, month, day etc. to Integer.parseInt(dayMonthYearArray[i]).

2. Extra level of nesting due to  String[] stringArray = userInput.split(" ", 2); being out of each possible command else if section.

Leads to worse readability due to greater amount unnecessary nesting, moreover the line String[] stringArray = userInput.split(" ", 2); is not clear what it does.

Add a comment line above the line to explain what it does and also put it into the else if sections for commands that need to use the line so as to remove one layer of nesting.