Implement Help Pop-up
It is difficult for the user to use wallE without knowing all the commands. Hardcoding the description in the fxml file is the fastest option to get this to feature up for now.

Future consideration is to have all the command formats and descriptions saved by Duke, and then rendered through a loop instead of hardcoding in the fxml file.