Update code to include tags
Currently, users are unable to organise their tasks in any way.

Let's include tags to allow for better organisation of tasks.

This can further be improved with tag filters, tag sorting etc.
With this, users can have a better view of their to-do list, i.e.
tagging the type of activity, the priority etc.