Create ViewSchedule Command
View command: returns all tasks recorded that fulfils the 2 criterias:
- The task is not marked.
- Depending on the task type, it is considered relevant.
- For Todo tasks, all tasks are considered relevant.
- For Deadline tasks, tasks are considered relevant if deadline date has not passed.
- For Event tasks, tasks are considered relevant if provided date falls within Event duration.

The View command provides a way to view tasks in the form of a schedule e.g., view the schedule for a specific date.

To use the command:
enter "view " followed by a date in the form of a formatted string in the given format: "yyyy-mm-dd".