Change exception type from Java-native types
Exceptions are currently mainly handled by NumberFormatException,
DateTimeParseException and TrackerBotException types, with the unchecked
exception IllegalStateException handling irrecoverable states in the
program.

The current method of handling makes it inconvenient to implement new
methods related to methods that throw the Java exception types, as
it is unclear whether the method should throw the native exception or
the specific Java exceptions. Furthermore, it complicates exception
handling downstream, as these specific scenarios need to be handled in
a way that users will be able to understand and rectify.

Therefore, we preemptively catch the Java exceptions and convert them
into a relevant TrackerBotException, that our methods downstream will be
able to handle generically.

This will allow us to program methods more generically in the long run,
and also let linters catch unhandled exceptions that we expect the user
to be able to create during the runtime of the program. In context of
the current branch, it lets us fix the assertions bug with an abstracted
method while also keeping specificity of the error to be thrown.