Add yml file for workflow enhancement
Currently, every git action that is done, is done manually.

Add a file that has a yml extension for the workflow commands.

This is tedious and time consuming for a programmer. Some times
the programmer might even forget some of these commands.
Some of these commands can be automated with a workflow aid.

An additional file helps to abstract away the logic of the program and
program aid tools.