Add assertions in convertInputDateAndTimeIntoLocalDateTime() method in Parser class.
User input day, month, year, hours and minutes are directly fed into the LocalDate.of() method.

The inputs go to the .of() method without being checked whether they are valid time inputs which might throw DateTimeException which gives a generic error message stating that there is a problem with the date time format in general.

Add assertions to check whether the time inputs are valid being passed into the .of() method.

This can give the user more specific error messages regarding their inputs which will be more useful to the user in correcting their user input.