Change gradle main method to target Launcher
Previously, the main method in gradle was assigned to that in the Hachi
class. As a result, the jar file created would use the main method in
Hachi, which did not have a graphical user interface.

Providing the user a graphical user interface will allow a more
streamlined and interactive user experience.

Let's change the build.gradle file to point to the main method in
Launcher, so the user can now use the GUI in the jar file.