Add sort feature to Jeoe
There is currently no sort feature.

Sometimes users want to see the list in a sorted manner for better
viewing.

Add a new sort command in Commands directory.
Add a sort method to Task manager.
Make Task objects comparable by extending comparable interface.
Add sort to the list of commands to the CommandParser.

Its done this way due to the abstraction of Jeoe.