feat:Add exit functionality after user enters bye command
If the user enters the bye command,
the MainWindow controller detects it,
then responds with the exit message.
It then uses JavaFX PauseTransition
to wait for 1.5 seconds before exiting.

This ensures that the window exits seamlessly
while also improving user experience,
as users could read the exit message
before the window exits.