Add assert statements
There are no checks for whether the TaskList and Storage objects passed into
commands' execute() methods are non-null, causing potential undefined
behaviour and errors.

Adding assert statements to the execute() methods allow us to verify that
the TaskList and Storage objects passed in are non-null.

Assert statements will cause the program to terminate if encountered, and can be disabled to avoid affecting application performance.