Create new JAR file with Duke.Launcher entry point
This commit adds a new JAR file to the project, with Duke.Launcher specified as the entry point.

This follows the change from a text bot where Duke.Duke was the entry point to now a GUI.

The JAR file is created to streamline the execution of the Duke application.

[Include any additional details or remarks here if necessary]

Signed-off-by: [Your Name] <youremail@example.com>