Add java assert in TaskList, Storage and DevyBot
TaskList, Storages and DevyBot have certain assumptions when adding new tasks,
saving files and returning the response.

Java assert helps to check for assumptions in the code.

I have identified areas to add assert such that the response to Main is never
empty. The new task created and being added to TaskList is never NULL. Storage
ensures that the directory and files are being created appropriately.