Increase flexibility of the find command.
find command only returns tasks with descriptions that match the keyword exactly.

This is not flexible as user might want to find a range of tasks that share similar letters.

Prioritises the tasks with descriptions that match more letters of the keyword always starting from the keyword first letter. Loop through the keyword substrings until left with only the first letter of the keyword. If the task description contains the substring of the keyword, then add the task into the result. Use a HashMap to track which tasks have already been added and which have not been added to the result array. If the task description contains the keyword entirely it will definitely be shown in the result, but if the description only contains a substring that is a part of the keyword, then it might not be added to the result if the limit has been reached. The limit can be set in the FindCommand.

Gives the user more flexibility when searching for a task.