Addition of Extension
There is no mechanism to check for duplicate task.

It is important to check for duplicate so that the task list will
correctly reflect the tasks that need to be completed.

A for loop is implemented to check if the new task is the same as any
of the existing tasks and if yes, the new task will not be added to the
task list.