Improve Code Quality
Further abstract methods which have more than 30 lines of code for
better readability. Then, split the codes into several methods where
applicable and necessary.

Use better names for variables and methods, names that self-define the
purpose of the variables and methods, for better readability.

Rephrase some of the JavaDoc comments so that they are more
understandable.