Improve abstraction of command classes and Parser
Command::execute handles reading, parsing and validating parameters.

This results in a very long execute function and has multiple layers of abstraction. This affects the readability of the code.

Add abstraction for loading, parsing and validating parameters that is called in the constructor.

When creating a new instance of the command, loading, parsing and validating can already be done instead of waiting until executing to check.