Implement ability to set and get reminders
There is no way to display a selection of tasks without specifying a specific date, hence users have to read through
the entire list of tasks to look for "important tasks".

Adding the ability to set and get reminders allows users to see just the important tasks (defined as tasks not complete
and are upcoming).

The option to set a reminder when creating a task hints to Atlas about the task's importance, thus users can be reminded
at the appropriate time.