Create feature to check duplicates
Add new feature to check duplicate when adding a new task.
The feature has three modes:
* ON: This will check duplicates for the description and exact timings of the task
* TEXT: This will only check duplicates in the descriptions. If two events have the same description but different timings, it will also register as a duplicate
* OFF: No duplicate check will be executed

Having duplicates checker prevents cluttering of the list so that it is cleaner to look at.
But, the different modes are useful if we want to make recurring events (It is okay to have duplicates)