Add an extension: D-Notes
Add a new feature which allows user to add a note for any task in the list. Users can also peek the notes for any task in the list by providing its index.

More specifically, sending "peekN" + index will show the notes for that task of the index, while sending "editN" + index + notes will update the note for the task to be the supplied note.

This feature is useful in the case when the user forget to include any details in the description of the task, and wish to add a notes to it.

This feature is also relevant to our TP as we also need a feature that allows users to add a bio/short intro of him/herself, the Notes feature can be easily modified to achieve the purpose.