Add duplicate task handling
Duplicate tasks is addable.

We do not want duplicate tasks to be addable by default.

Lets make the the application reply to the user with a message
saying they tried to add a duplicate description upon
attempting to add duplicate description.

Allows the user to know if they have accidentally added a
overlapping task description.