Fix coding standard violations in classes
The code in classes had several violations of the Java coding standard
as pointed out in the comments on the pull request. This commit
addresses those issues and ensures that the code now adheres to the
prescribed coding standard.

Changes made:
- Moved println statements to Ui class.
- Renamed MondayExceptions to MondayException.
- Add a TaskList variable to Parser.
- Renamed unMarkTask to markAsUndone in Ui class.
- Renamed Storage to TaskArrayListStorage.
- Removed redundant try block in the function save.
- Renamed addToTask to addToList
- Renamed displayList to toString

This commit improves code readability and maintainability by adhering
to the recommended coding conventions.