Add ability to sort deadlines and events
When there are many different types of tasks in the main list, it is
difficult to see which deadlines and events are coming up.

Allow users to sort deadlines and events by date so that they can see
which ones are approaching.

This helps to improve the usability of the bot.