Add help command
When the user uses the app for the first time, he or she will not know
what the available commands are.

Adding a new command called "help" will allow the user to view all the
available commands.

Let's create a new command class for the help command, which implements
the command interface, as this will allow us to practice good OOP
principles. We should also add support for parsing this new command and
prompt the user to use this command in the event where they type in an
invalid command.