overhauled the entire Duchess system to focus on using callback handlers instead.
Previously, Duchess and DuchessGUI were both separate classes each with
their own versions of the same method to handle the different cases of
CLI Duchess and GUI Duchess. However, with this change, they now both
rely on callback handlers to differentiate the two cases.

This means that Duchess and DuchessGUI no longer need to be separate
classes, but can each both have a Duchess instance running the same
commands, the only difference being the callback handler.