Add assertions
Without making assertion checks, ensuring the correctness of the assumptions made in the application can be quite challenging.

Assert reassures the assumption made in the program.

To check the command assumption for command exceution functions, assertions can be used to reassure that the command used in each function matches what is needed.

Assertion can also be useful to reassure the index input of task edit functions with the assumption that the input index is a valid index.

Utilizing the Java assert statement serves as an effective method for conducting runtime validations within a program. This enables the early detection of unforeseen issues and failures during the development phase.