Add help command
A user currently does not have a way to view the functions
available and how to use them.

A user will be unsure how to use the program, not knowing the
commands available.

Lets add a "help" command that allows users to view the commands
available to them.

Using a help command is a simple way for users to easily
understand how to use the program.