Change mass delimiter to single character
The Mass command uses double colons "::" to delimit each subsequent
command of the same subtype.

For a command designed to shorthand mass operations, this is very
cumbersome to have to SHIFT+";" twice. Additionally, there is no
real reason to use double colons in the code, as colons and
semicolons do not have a special meaning in the parser.

Therefore, we adjust the delimiter to use a single semicolon, which
should make the process less cumbersome.