Added JUnit test methods for deadline, event and todo.
There are several test cases to check the functionality of task classes.
The test checks for the correct string representation after the given
input is entered.

Include javadoc for all methods.

Added javadoc for all methods for better documentation and readability.

Added Assertions

Added assertions to check for null or invalid variable representations
to debug.