Implement assertions in Crusader
Many parts of Crusader code make implicit assumptions about
conditions that are required to let them execute properly.
This is unsafe as these conditions may no longer be true as
the project evolves.

Using assertions allows us to make such assumptions explicit, so that
contributors are better aware of the required conditions for our project
to behave as expected.

Let's allow for the use of assertions in the code, as well as begin to
use assertions throughout the project.