Improve coding standard of program
The code has many areas which violate the coding standard such as
heavily nested loops and overly long methods which affect the
readability of the code.

Improving the coding standard by rewriting overly long methods to
shorter ones as well as avoiding nested loops can prevent issues in
the future where the code is hard to understand and read.

Let's improve the coding standard by rewriting parts of the code
which violate the coding standard.