Move logic of commands into individual Command classes
Execution of commands is being done by TaskList class instead of
command classes.

It may be difficult to ascertain any bugs with Command logic or TaskList
logic if they are not separated.

The logic of commands being executed will be handled by the command
classes instead of the TaskList class so TaskList will only have methods
relevant to the TaskList.

With proper separation of behaviors into relevant classes, the code will
be clearer.