Add a User Guide to the project
The code will now have a user guide for users
to understand how to use the application.

The lack of user guide would make it difficult for users
to know how to use the application.

Including the user guide allows users to have a
better idea on how to use the application.

A detailed user guide is preferred to ensure users understand
how to use the essential features of the application.